requirements.txt
dj-database-url
Django
django-allauth
django-ckeditor
django-cors-headers
django-countries
django-crispy-forms
django-debug-toolbar
django-extensions
django-import-export
django-storages[google]
django-taggit
djangorestframework
gunicorn
numpy
pandas
Pillow
psycopg2-binary
pycco
